Moved mailmerge to src-layout
