+++ src/mailmerge/__init__.py
@@ -1,0 +1,4 @@
+__version__ = "0.8.3"
+
+from .mailmerge import NAMESPACES as NAMESPACES
+from .mailmerge import MailMerge as MailMerge
