+++ tox.ini
@@ -7,11 +7,11 @@ commands=python -m unittest discover
 [testenv:flake8]
 commands=
     flake8 tests --extend-exclude=tests/old/
-    flake8 mailmerge.py
+    flake8 src/mailmerge/
 deps=flake8
 
 [flake8]
-ignore=W999,E203,F841,W503,E501
+ignore=W999,E203,F841,W503,E501,F401
 max-line-length=120
 
 [testenv:lint]
